@phohensee 2023-03-02
The linux-x86 pre-submit test failure is caused by a test using -XX:+UseCompressedClassPointers, which is an invalid switch for 32-bit JVMs.

The linux-cross-compile pre-submit test failure is a compile-time failure in src/hotspot/cpu/arm/interpreterRT_arm.cpp, which latter is not touched by this patch.